requirement.txt
ansible-base==2.10.8
apache-libcloud==3.2.0
argcomplete==1.8.1
asgiref==3.7.2
Babel==2.8.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.10
django-cors-headers==4.1.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.1.0
filelock==3.6.0
finnhub-python==2.4.18
Flask==2.2.3
gunicorn==20.1.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
MarkupSafe==2.1.2
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
oauthlib==3.2.0
packaging==21.3
platformdirs==2.5.1
pycryptodomex==3.11.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pykerberos==1.1.14
pymongo==4.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
pytz==2022.1
pywinrm==0.3.0
PyYAML==5.4.1
requests==2.25.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
selinux==3.3
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
supervisor==4.2.1
systemd-python==234
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
Werkzeug==2.2.3
xmltodict==0.12.0
zipp==1.0.0
